move out query generation code
