+++ mysql-store.js
@@ -1,18 +1,16 @@
 'use strict'
+
 var Assert = require('assert')
 var _ = require('lodash')
 var MySQL = require('mysql')
 var UUID = require('node-uuid')
 var DefaultConfig = require('./default_config.json')
+var QueryBuilder = require('./query-builder')
 
 var Eraro = require('eraro')({
   package: 'mysql'
 })
 
-var OBJECT_TYPE = 'o'
-var ARRAY_TYPE = 'a'
-var DATE_TYPE = 'd'
-var SENECA_TYPE_COLUMN = 'seneca'
 var storeName = 'mysql-store'
 
 module.exports = function (options) {
@@ -189,13 +187,13 @@ module.exports = function (options) {
         }
       }
 
-      var entp = makeentp(ent)
+      var entp = QueryBuilder.makeentp(ent)
 
       if (update) {
-        query = 'UPDATE ' + tablename(ent) + ' SET ? WHERE id=\'' + entp.id + '\''
+        query = 'UPDATE ' + QueryBuilder.tablename(ent) + ' SET ? WHERE id=\'' + entp.id + '\''
       }
       else {
-        query = 'INSERT INTO ' + tablename(ent) + ' SET ?'
+        query = 'INSERT INTO ' + QueryBuilder.tablename(ent) + ' SET ?'
       }
 
       internals.connectionPool.query(query, entp, function (err, result) {
@@ -246,7 +244,7 @@ module.exports = function (options) {
             return done(err)
           }
 
-          var ent = makeent(qent, res[0])
+          var ent = QueryBuilder.makeent(qent, res[0])
 
           seneca.log(args.tag$, 'load', ent)
 
@@ -278,7 +276,7 @@ module.exports = function (options) {
 
       var qent = args.qent
       var q = args.q
-      var queryfunc = makequeryfunc(qent, q, internals.connectionPool)
+      var queryfunc = QueryBuilder.makequeryfunc(qent, q, internals.connectionPool)
 
       queryfunc(function (err, results) {
         if (err) {
@@ -287,7 +285,7 @@ module.exports = function (options) {
 
         var list = []
         results.forEach(function (row) {
-          var ent = makeent(qent, row)
+          var ent = QueryBuilder.makeent(qent, row)
           list.push(ent)
         })
         cb(null, list)
@@ -328,7 +326,7 @@ module.exports = function (options) {
       }
 
       function executeRemove (args, row) {
-        var query = deletestm(qent, q, internals.connectionPool)
+        var query = QueryBuilder.deletestm(qent, q, internals.connectionPool)
 
         internals.connectionPool.query(query, function (err, result) {
           if (err) {
@@ -377,210 +375,30 @@ module.exports = function (options) {
     var qent = args.qent
     q.limit$ = 1
 
-    var query = selectstm(qent, q, internals.connectionPool)
+    var query = QueryBuilder.selectstm(qent, q, internals.connectionPool)
     return done(null, {query: query})
   })
 
   seneca.add({role: storeName, hook: 'save'}, function (args, done) {
+    var ent = args.ent
+    var update = !!ent.id
+    var query
+    var entp = QueryBuilder.makeentp(ent)
 
+    if (update) {
+      query = 'UPDATE ' + QueryBuilder.tablename(ent) + ' SET ? WHERE id=\'' + entp.id + '\''
+    }
+    else {
+      query = 'INSERT INTO ' + QueryBuilder.tablename(ent) + ' SET ?'
+    }
+
+    if (update) {
+      return done(null, {query: query, operation: 'save/update'})
+    }
+    else {
+      return done(null, {query: query, operation: 'save/insert'})
+    }
   })
 
   return {name: store.name, tag: meta.tag}
-}
-
-var fixquery = function (qent, q) {
-  var qq = {}
-  for (var qp in q) {
-    if (!qp.match(/\$$/)) {
-      qq[qp] = q[qp]
-    }
-  }
-  return qq
-}
-
-
-var whereargs = function (qent, q) {
-  var w = {}
-  var qok = fixquery(qent, q)
-
-  for (var p in qok) {
-    w[p] = qok[p]
-  }
-  return w
-}
-
-
-var selectstm = function (qent, q, connection) {
-  var table = tablename(qent)
-  var params = []
-  var w = whereargs(makeentp(qent), q)
-  var wherestr = ''
-
-  if (!_.isEmpty(w)) {
-    for (var param in w) {
-      params.push(param + ' = ' + connection.escape(w[param]))
-    }
-    wherestr = ' WHERE ' + params.join(' AND ')
-  }
-
-  var mq = metaquery(qent, q)
-  var metastr = ' ' + mq.join(' ')
-
-  return 'SELECT * FROM ' + table + wherestr + metastr
-}
-
-
-var tablename = function (entity) {
-  var canon = entity.canon$({object: true})
-  return (canon.base ? canon.base + '_' : '') + canon.name
-}
-
-
-var makeentp = function (ent) {
-  var entp = {}
-  var fields = ent.fields$()
-  var type = {}
-
-  fields.forEach(function (field) {
-    if (_.isArray(ent[field])) {
-      type[field] = ARRAY_TYPE
-    }
-    else if (!_.isDate(ent[field]) && _.isObject(ent[field])) {
-      type[field] = OBJECT_TYPE
-    }
-
-    if (!_.isDate(ent[field]) && _.isObject(ent[field])) {
-      entp[field] = JSON.stringify(ent[field])
-    }
-    else {
-      entp[field] = ent[field]
-    }
-  })
-
-  if (!_.isEmpty(type)) {
-    entp[SENECA_TYPE_COLUMN] = JSON.stringify(type)
-  }
-  return entp
-}
-
-
-var makeent = function (ent, row) {
-  if (!row) {
-    return null
-  }
-  var entp
-  var fields = _.keys(row)
-  var senecatype = {}
-
-  if (!_.isUndefined(row[SENECA_TYPE_COLUMN]) && !_.isNull(row[SENECA_TYPE_COLUMN])) {
-    senecatype = JSON.parse(row[SENECA_TYPE_COLUMN])
-  }
-
-  if (!_.isUndefined(ent) && !_.isUndefined(row)) {
-    entp = {}
-    fields.forEach(function (field) {
-      if (SENECA_TYPE_COLUMN !== field) {
-        if (_.isUndefined(senecatype[field])) {
-          entp[field] = row[field]
-        }
-        else if (senecatype[field] === OBJECT_TYPE) {
-          entp[field] = JSON.parse(row[field])
-        }
-        else if (senecatype[field] === ARRAY_TYPE) {
-          entp[field] = JSON.parse(row[field])
-        }
-        else if (senecatype[field] === DATE_TYPE) {
-          entp[field] = new Date(row[field])
-        }
-      }
-    })
-  }
-  return ent.make$(entp)
-}
-
-
-var metaquery = function (qent, q) {
-  var mq = []
-
-  if (q.sort$) {
-    for (var sf in q.sort$) break
-    var sd = q.sort$[sf] < 0 ? 'DESC' : 'ASC'
-    mq.push('ORDER BY ' + sf + ' ' + sd)
-  }
-
-  if (q.limit$) {
-    mq.push('LIMIT ' + (Number(q.limit$) || 0))
-  }
-
-  if (q.skip$) {
-    mq.push('OFFSET ' + (Number(q.skip$) || 0))
-  }
-
-  return mq
-}
-
-
-function makequeryfunc (qent, q, connection) {
-  var qf
-  if (_.isArray(q)) {
-    if (q.native$) {
-      qf = function (cb) {
-        var args = q.concat([cb])
-        connection.query.apply(connection, args)
-      }
-      qf.q = q
-    }
-    else {
-      qf = function (cb) {
-        connection.query(q[0], _.tail(q), cb)
-      }
-      qf.q = {q: q[0], v: _.tail(q)}
-    }
-  }
-  else if (_.isObject(q)) {
-    if (q.native$) {
-      var nq = _.clone(q)
-      delete nq.native$
-      qf = function (cb) {
-        connection.query(nq, cb)
-      }
-      qf.q = nq
-    }
-    else {
-      var query = selectstm(qent, q, connection)
-      qf = function (cb) {
-        connection.query(query, cb)
-      }
-      qf.q = query
-    }
-  }
-  else {
-    qf = function (cb) {
-      connection.query(q, cb)
-    }
-    qf.q = q
-  }
-
-  return qf
-}
-
-
-var deletestm = function (qent, q, connection) {
-  var table = tablename(qent)
-  var params = []
-  var w = whereargs(makeentp(qent), q)
-  var wherestr = ''
-
-  if (!_.isEmpty(w)) {
-    for (var param in w) {
-      params.push(param + ' = ' + connection.escape(w[param]))
-    }
-    wherestr = ' WHERE ' + params.join(' AND ')
-  }
-
-  var limistr = ''
-  if (!q.all$) {
-    limistr = ' LIMIT 1'
-  }
-  return 'DELETE FROM ' + table + wherestr + limistr
 }
